chore: stop using coveralls because of a GPL-licensed transitive dep
Our `coveralls` development dependency imports the `term-ansicolor`
dependency. This dependency is licensed under GPLv2 which we do not want
to use as it would make our users' projects subject to GPL limitations.

To counter that, we tried enforce the `term-ansicolor` dependency to
`1.6.0`.  As of this version, this dependency is using an Apache 2.0
license which is compatible with both non-GPL and GPL users' projects.

However, because `term-ansicolor` as of `1.6.0` is not compatible with
Ruby 1.x versions, the last solution is to completely get rid of
`coveralls`.

[changelog]