refactor: replace Ruby 2.3>= hash comparison operator with Helpers.hash_includes_subset?
The `<=` and `>=` Hash operators have been introduced in Ruby 2.3 and
hence, cannot be used in our codebase which supports at least Ruby 2.2.
This commit introduces a new `Helpers.hash_includes_subset?` function to
replace the operator we used in the `wait` method of the
`UpdateApiKeyResponse` class.

For more informations, see the `Hash` section of this NEWS note from
Ruby 2.3.0:
https://github.com/ruby/ruby/blob/v2_3_0/NEWS#core-classes-updates-outstanding-ones-only-