fix(logging): pass user configured logger to temp client (#480)
The `replace_all_objects` command creates a temporary client which does
not have the some configuration as one constructed by users of the
library. This change adds an optional parameter when constructing an
index for the `logger` and allows that to be passed in to the temporary
client. This change also ensures that any `Index` built through the
`SearchClient#init_index` call gets the logger configured on the client.
This will ensure logging is done in a consistent manner across clients
instantiated by the library.